frontend/requirements.txt
streamlit==1.40.2
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
